move PatientSnapshot
